analysis/requirements.txt
pyspark
python-dotenv
pymysql
